Development ss16

+++ latex/day4/problems/listings/functional_intro_example.cpp
@@ -1,7 +1,7 @@
-std::vector<int> numbers = { 1, 2, 3, 4, 5 };
+std::vector<double> numbers = { 1, 2, 3, 4, 5 };
 
-for (std::vector<int>::iterator it = numbers.begin(); it != numbers.end(); ++it) {
-    *it = square(*it); // squaring element pointed to by it
+for (std::vector<double>::iterator it = numbers.begin(); it != numbers.end(); ++it) {
+    *it = square(*it); // squaring element podoubleed to by it
 }
 
 // numbers now [ 1, 4, 9, 16, 25 ]
